Began the results, added the total returns to script, see message
Added the calculation for the total returns to the R file, also
added a table comparing both trading stats to the paper. Should make it
easier to write about. Maybe I should fix the rownames as it looks pretty
ugly right now. Also wrote a piece of the Results section.
Added all the graphics, the latex header logo etc, too